requirements.txt
attrs==19.3.0
aws-cdk.assets==1.31.0
aws-cdk.aws-apigateway==1.31.0
aws-cdk.aws-applicationautoscaling==1.31.0
aws-cdk.aws-autoscaling==1.31.0
aws-cdk.aws-autoscaling-common==1.31.0
aws-cdk.aws-autoscaling-hooktargets==1.31.0
aws-cdk.aws-batch==1.31.0
aws-cdk.aws-certificatemanager==1.31.0
aws-cdk.aws-cloudformation==1.31.0
aws-cdk.aws-cloudfront==1.31.0
aws-cdk.aws-cloudwatch==1.31.0
aws-cdk.aws-codebuild==1.31.0
aws-cdk.aws-codecommit==1.31.0
aws-cdk.aws-codepipeline==1.31.0
aws-cdk.aws-ec2==1.31.0
aws-cdk.aws-ecr==1.31.0
aws-cdk.aws-ecr-assets==1.31.0
aws-cdk.aws-ecs==1.31.0
aws-cdk.aws-elasticloadbalancing==1.31.0
aws-cdk.aws-elasticloadbalancingv2==1.31.0
aws-cdk.aws-events==1.31.0
aws-cdk.aws-events-targets==1.31.0
aws-cdk.aws-iam==1.31.0
aws-cdk.aws-kms==1.31.0
aws-cdk.aws-lambda==1.31.0
aws-cdk.aws-logs==1.31.0
aws-cdk.aws-route53==1.31.0
aws-cdk.aws-route53-targets==1.31.0
aws-cdk.aws-s3==1.31.0
aws-cdk.aws-s3-assets==1.31.0
aws-cdk.aws-sam==1.31.0
aws-cdk.aws-secretsmanager==1.31.0
aws-cdk.aws-servicediscovery==1.31.0
aws-cdk.aws-sns==1.31.0
aws-cdk.aws-sns-subscriptions==1.31.0
aws-cdk.aws-sqs==1.31.0
aws-cdk.aws-ssm==1.31.0
aws-cdk.aws-stepfunctions==1.31.0
aws-cdk.core==1.31.0
aws-cdk.cx-api==1.31.0
aws-cdk.region-info==1.31.0
boto3==1.12.43
botocore==1.15.43
cattrs==1.0.0
certifi==2019.11.28
chardet==3.0.4
constructs==2.0.1
docutils==0.15.2
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.4.0
jmespath==0.9.5
jsii==1.1.0
publication==0.0.3
python-dateutil==2.8.1
requests==2.23.0
serverless_py==0.0.1
s3transfer==0.3.3
six==1.14.0
typing-extensions==3.7.4.1
urllib3==1.25.8
zipp==3.1.0
